+++ extensions/now-playing/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- Added configurable menu bar display presets and a custom template mode.
 - Initial release.